Enable YAML targets
Add a second QFYAML configuration object (yml_anchored) so that we
can keep a list of variables that point to targets.  Also, we now
move the creation of variables that point to targets out of
parse_line and into QFYAML_read_line.

Corrected an issue with the species database: added MolecRatio for
ACET and also made "NEEDED" into a comment to avoid parsing issues.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>