Now return an error code all the way up to the calling routine
Prior to this commit, the Handle_Error function would stop with an
error.  This can have a detrimental impact when running in an MPI
environment.

A more robust way of trapping errors is to pass back a return code
(RC) variable from each routine.  This can be passed up to the main
program, where the run can be gracefully ended if an error occurs.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>