Add an error trap that returns if we can't open a species db file
In src/qfyaml_mod.F90 we have now added an error trap that calls
Handle_Error to print a message and return if we cannot find the
given species database file on disk.  This error can propagage all
the way up tot he main program, where it can be shut down gracefully.

Also updated the GEOS-Chem species_database*yml files as of 5/5/20.

Lastly, fixed a typo in README.md so that we now point to the proper
license agreement.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>