Bug fix: Now fix parsing of categories more than 2 indent levels prior
It was discovered that this YAML markup was not being parsed properly
in qfyaml 0.3.2:

   operations:
     transport:
       passive_species:
         CH3ITracer:
           long_name: Methyl_iodide
           mol_wt_in_g: 142.0
           lifetime_in_s: 4.32e5
           default_bkg_conc_in_vv: 1.0e-20
     wet_deposition:
       activate: true

The wet_deposition category was given the wrong indentation depth,
and as such, operations:wet_deposition:activate was parsed as FALSE
but should have been TRUE.

This issue is now fixed and we will release this as qfyaml 0.3.0.

NOTE: QFYAML works best if a constant indentation (e.g. always 2 spaces
for each level) is used.  You can use an editor such as Emacs to format
your YAML input files accordingly.

Also updated the test programs to test this behavior.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>